requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
alembic==1.14.0
pydantic[email]==2.10.6
pydantic-settings==2.7.1
python-multipart==0.0.20
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
python-dotenv==1.0.1
aiosmtplib==3.0.2
jinja2==3.1.5
pytest==8.3.4
httpx==0.28.1
